requirements.txt
langchain==0.1.0
langchain-openai==0.1.0
streamlit==1.28.1
python-dotenv==1.0.0
openai==1.3.0
